requirements.txt
alembic==1.9.2
autopep8==2.0.1
boto3==1.26.67
botocore==1.29.67
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask==2.2.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.2
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==0.21.1
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==2.0.1
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
